Move Node inline functions from ShadowRoot.h to NodeInlines.h https://bugs.webkit.org/show_bug.cgi?id=297017
Reviewed by Wenson Hsieh and Aditya Keerthi.

Moved Node inline functions defined in ShadowRoot.h to NodeInlines.h.

Canonical link: https://commits.webkit.org/298373@main
